+++ src/lib.rs
@@ -230,6 +230,7 @@ pub use shrev::ReaderId;
 pub use shred::AsyncDispatcher;
 
 pub use changeset::ChangeSet;
-pub use storage::{DenseVecStorage, FlaggedStorage, HashMapStorage, InsertedFlag, ModifiedFlag, 
-                  NullStorage, ReadStorage, RemovedFlag, Storage, Tracked, VecStorage, WriteStorage};
-pub use world::{Component, Entities, Entity, EntityBuilder, LazyUpdate, World};
+pub use storage::{DenseVecStorage, FlaggedStorage, HashMapStorage, InsertedFlag, ModifiedFlag,
+                  NullStorage, ReadStorage, RemovedFlag, Storage, Tracked, VecStorage,
+                  WriteStorage};
+pub use world::{Builder, Component, Entities, Entity, EntityBuilder, LazyUpdate, World};